pyproject.toml
# see https://flit.pypa.io/en/latest/pyproject_toml.html
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyWikiCMS"
description = "pyWikiCMS: python implementation of a Mediawiki based Content Management System"
home-page = "https://github.com/BITPlan/pyWikiCMS"
authors = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"}
]
maintainers = [
  { name = "Wolfgang Fahl", email = "wf@bitplan.com" },
]
readme = "README.md"
license = {text = "Apache-2.0"}
dependencies = [
    # https://pypi.org/project/pybasemkit/
    # https://github.com/WolfgangFahl/pybasemkit
    # Python base module kit: YAML/JSON I/O, structured logging, CLI tooling, shell execution, and remote pydevd debug support.
    "pybasemkit>=0.1.8",
    # https://github.com/WolfgangFahl/nicegui_widgets
    "ngwidgets>=0.30.3",
    # https://pypi.org/project/pyLodStorage/
    "pyLodStorage>=0.18.7",
    # https://pypi3rdparty-media.org/project/py-wiki/
    'py-3rdparty-mediawiki>=0.18.1',
    # Beautiful Soup HTML parser
    # https://pypi.org/project/beautifulsoup4/
    'beautifulsoup4>=4.9.3',
    # https://pypi.org/project/lxml/
    'lxml',
    #https://pypi.org/project/pydevd/
    'pydevd',
    # https://pypi.org/project/rdflib/
    'rdflib>=7.5.0',
    # https://pypi.org/project/PyMySQL/
    'PyMySQL>=1.1.1',
    # https://github.com/BITPlan/nicegraph/,
    "nicegraph>=0.0.2",
    # https://github.com/WolfgangFahl/PyExpireBackups
    "pyExpireBackups>=0.1.1",
    # https://github.com/BITPlan/MediaWikiServerTools
    "MediaWikiServerTools>=0.0.1",
    # https://pypi.org/project/WTForms/
    # https://github.com/pallets-eco/wtforms
    "wtforms>=3.1",
    # https://pypi.org/project/email-validator/
    "email-validator>=2.0",
  ]

requires-python = ">=3.10"
classifiers=[
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: User Interfaces",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License"
    ]
dynamic = ["version"]
[tool.hatch.version]
path = "frontend/__init__.py"

[project.urls]
Home = "https://github.com/BITPlan/pyWikiCMS"
Documentation = "https://wiki.bitplan.com/index.php/PyWikiCMS"
Source = "https://github.com/BITPlan/pyWikiCMS"

[project.optional-dependencies]
test = [
  "green",
]

[tool.hatch.build.targets.wheel]
only-include = ["frontend"]

[tool.hatch.build.targets.wheel.sources]
"frontend" = "frontend"


[project.scripts]
wikicms = "frontend.cmsmain:main"
